Refactor into aspi package, unified CLI entrypoint
Split the codebase into a proper Python package:

- aspi/ldcs.py      — Lark grammar + base LDCS transformer
- aspi/asp.py       — ASP/clingo backend (was aspi.py)
- aspi/prolog/      — Prolog backend (was prolog.py, ~2000 lines)
  - ast.py          — PTerm/PGoal dataclasses
  - analysis.py     — variable extraction, dependency tracking
  - render.py       — term/goal/body rendering with when/2 autowrap
  - macros.py       — _MacroExpander for user-defined macros
  - compiler.py     — PrologLDCS transformer (LDCS → Prolog AST)
  - engine.py       — PrologEngine (subprocess swipl)
  - formatting.py   — result parsing and sorting
  - repl.py         — PrologASPI hybrid REPL
- aspi/__main__.py  — unified CLI: python -m aspi

Old top-level aspi.py, prolog.py, ldcs.py deleted (no compat shims).
Obsolete environment.yml (conda/Python 3.8) removed.
All test imports updated. pytest-console-scripts dependency dropped.
