+++ aspi/__init__.py
@@ -1,0 +1,1 @@
+"""aspi — Answer Set Programming, Interactively."""
