--- environment.yml
@@ -1,18 +1,0 @@
-# https://docs.conda.io/en/latest/miniconda.html
-# conda env create -f environment.yml
-name: aspi
-channels:
-  - conda-forge
-  - potassco
-dependencies:
-  - python=3.8
-  - clingo
-  - flake8
-  - lark-parser
-  - mypy
-  - sh
-  - pytest
-  - pytest-console-scripts
-  - pip
-  - pip:
-    - flake8-mypy
